backend/requirements.txt
django
djangorestframework
django-cors-headers
django-stubs
djangorestframework-stubs